+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmMapInternal.kt
@@ -145,7 +145,8 @@ internal fun <K, V : BaseRealmObject> ManagedRealmMap<K, V?>.query(
  */
 internal interface MapOperator<K, V> : CollectionOperator<V, RealmMapPointer> {
 
-    // Modification counter used to detect concurrent writes from the iterator
+    // Modification counter used to detect concurrent writes from the iterator, taken from Java's
+    // AbstractList implementation
     var modCount: Int
     val keyConverter: RealmValueConverter<K>
     override val nativePointer: RealmMapPointer
@@ -845,18 +846,16 @@ internal abstract class RealmMapGenericIterator<K, T>(
     abstract fun getNext(position: Int): T
 
     override fun hasNext(): Boolean {
-        operator.realmReference.checkClosed()
         checkConcurrentModification()
 
         return cursor < operator.size
     }
 
     override fun remove() {
-        operator.realmReference.checkClosed()
         checkConcurrentModification()
 
         if (operator.size == 0) {
-            throw NoSuchElementException("Could not remove last element returned by the iterator: set is empty.")
+            throw NoSuchElementException("Could not remove last element returned by the iterator: dictionary is empty.")
         }
         if (lastReturned < 0) {
             throw IllegalStateException("Could not remove last element returned by the iterator: iterator never returned an element.")
@@ -880,7 +879,6 @@ internal abstract class RealmMapGenericIterator<K, T>(
     }
 
     override fun next(): T {
-        operator.realmReference.checkClosed()
         checkConcurrentModification()
 
         val position = cursor